% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimator.R
\name{em.cov.bs}
\alias{em.cov.bs}
\title{Bootstrap covariance estimator of EM point estimator}
\usage{
em.cov.bs(
  theta.em,
  counts,
  m = 10000,
  em.eps = 1e-06,
  bs.eps = 1e-04,
  debug = F
)
}
\arguments{
\item{theta.em}{An EM point estimator of theta given observed counts}

\item{counts}{observed count data (n0,n1,...,n16)}

\item{m}{maximum bootstrap replicates}

\item{em.eps}{EM algorithm epsilon stopping condition}

\item{bs.eps}{Bootstrap covariance epsilon stopping condition}

\item{debug}{whether to print out debugging info while running}
}
\description{
Bootstrap covariance estimator of EM point estimator
}
\keyword{EM}
\keyword{algorithm}
