% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimator.R
\name{em.cov.info}
\alias{em.cov.info}
\title{Covariance matrix of EM point estimator based on the observed information matrix}
\usage{
em.cov.info(theta.em, counts)
}
\arguments{
\item{theta.em}{An EM point estimator of theta given observed responses}

\item{data}{Observed sample of responses}
}
\description{
Covariance matrix of EM point estimator based on the observed information matrix
}
