% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimator.R
\name{em.cov_estimator.bootstrap}
\alias{em.cov_estimator.bootstrap}
\title{Bootstrap covariance estimator of EM point estimator}
\usage{
em.cov_estimator.bootstrap(theta.em, counts, m = 10000, debug = T)
}
\arguments{
\item{theta.em}{An EM point estimator of theta given observed counts}

\item{counts}{observed count data (n0,n1,...,n16)}

\item{m}{bootstrap replicates}

\item{debug}{whether to print out debugging info while running}
}
\description{
Bootstrap covariance estimator of EM point estimator
}
\keyword{EM}
\keyword{algorithm}
