% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimator.R
\name{em.cov_estimator.observed_info}
\alias{em.cov_estimator.observed_info}
\title{Covariance matrix of EM point estimator based on the observed information matrix}
\usage{
em.cov_estimator.observed_info(theta.em, data)
}
\arguments{
\item{theta.em}{An EM point estimator of theta given observed responses}

\item{data}{Observed sample of responses}
}
\description{
Covariance matrix of EM point estimator based on the observed information matrix
}
\keyword{covariance}
\keyword{estimator}
