% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_estimator.R
\name{em.estimator}
\alias{em.estimator}
\title{EM algorithm}
\usage{
em.estimator(theta, counts, eps = 1e-06, debug = T)
}
\arguments{
\item{theta}{a starting guess for theta = (alpha,beta,mu,lambda)}

\item{counts}{observed count data (n0,n1,...,n16)}

\item{eps}{stopping condition}

\item{debug}{whether to print out debugging info while running}
}
\description{
EM algorithm estimator for problem 4.2
}
