% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimator.R
\name{em.observed_info}
\alias{em.observed_info}
\title{Observed information of EM point estimator based on an observed sample}
\usage{
em.observed_info(theta.em, counts)
}
\arguments{
\item{theta.em}{An EM point estimator of theta given observed counts}

\item{data}{Observed sample of responses}
}
\description{
Observed information of EM point estimator based on an observed sample
}
