% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_fn.R
\name{em.pdf}
\alias{em.pdf}
\title{probability density function (pdf) for problem 4.2}
\usage{
em.pdf(x, theta)
}
\arguments{
\item{x}{density evaluated at x given index theta}

\item{theta}{index of indexed family of density functions, theta = (alpha,beta,mu,lambda)}
}
\description{
probability density function (pdf) for problem 4.2
}
\keyword{density}
