% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimator.R
\name{em.sd.bs}
\alias{em.sd.bs}
\title{Standard error of EM point estimator based on Bootstrapping sample covariance}
\usage{
em.sd.bs(
  theta.em,
  counts,
  m = 10000,
  em.eps = 1e-06,
  bs.eps = 1e-04,
  debug = F
)
}
\arguments{
\item{theta.em}{An EM point estimator of theta given observed responses}

\item{count}{Observed sample of counts}
}
\description{
Standard error of EM point estimator based on Bootstrapping sample covariance
}
