% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimator.R
\name{em.sd.info}
\alias{em.sd.info}
\title{Standard error of EM point estimator based on the observed information matrix}
\usage{
em.sd.info(theta.em, counts)
}
\arguments{
\item{theta.em}{An EM point estimator of theta given observed responses}

\item{count}{Observed sample of counts}
}
\description{
Standard error of EM point estimator based on the observed information matrix
}
