% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_convert.R
\name{em.counts_to_responses}
\alias{em.counts_to_responses}
\title{Convert counts to individual responses}
\usage{
em.counts_to_responses(counts)
}
\arguments{
\item{counts}{Count data to convert}
}
\description{
There are many ways to represent the same data. In problem 4.2, the data
is given in the form of counts, e.g., counts\link{j} represents the number
of people who responded with j risky sexual encounters.
The more natural form of this data is as a vector of indivual responses,
e.g., x\link{k} is the k-th persons response to the number of risky sexual
encounters.
}
\keyword{counts}
\keyword{multinomial}
\keyword{response}
