% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_fn.R
\name{em.loglike}
\alias{em.loglike}
\title{log-likelihood function for problem 4.2}
\usage{
em.loglike(theta, data)
}
\arguments{
\item{theta}{evaluated at theta = (alpha,beta,mu,lambda)}

\item{data}{observed response data}
}
\description{
log-likelihood function for problem 4.2
}
