% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_convert.R
\name{em.responses_to_counts}
\alias{em.responses_to_counts}
\title{Convert individual response data to count data}
\usage{
em.responses_to_counts(data)
}
\arguments{
\item{data}{response data}
}
\value{
response data converted to count data
}
\description{
This is the inverse of em.counts_to_responses, e.g.,
em.responses_to_counts(em.counts_to_responses(counts)) == counts
and
em.counts_to_responses(em.responses_to_counts(data)) == data.
}
