% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_convert.R
\name{em.responses_to_counts}
\alias{em.responses_to_counts}
\title{Convert individual responses to count data}
\usage{
em.responses_to_counts(data)
}
\arguments{
\item{data}{Vector of responses
export}
}
\description{
Convert individual responses to count data
}
